section004/requirement.txt
asgiref==3.8.1
Django==5.1.6
django-filter==25.1
djangorestframework==3.15.2
gitdb==4.0.11
GitPython==3.1.41
Markdown==3.7
setuptools==69.0.3
smmap==5.0.1
sqlparse==0.5.3
